feat: assemble partial transaction
